+++ diive/__init__.py
@@ -41,6 +41,8 @@ from diive.pkgs.corrections.setto import setto_threshold as setto_threshold
 from diive.pkgs.corrections.offsetcorrection import MeasurementOffsetFromReplicate as MeasurementOffsetFromReplicate
 from diive.pkgs.corrections.offsetcorrection import remove_relativehumidity_offset as remove_relativehumidity_offset
 from diive.pkgs.corrections.offsetcorrection import remove_radiation_zero_offset as remove_radiation_zero_offset
+from diive.pkgs.corrections.offsetcorrection import WindDirOffset as WindDirOffset
+from diive.pkgs.corrections.offsetcorrection import WindDirOffset as wind_dir_offset
 from diive.pkgs.createvar.conversions import et_from_le as et_from_le
 from diive.pkgs.echires.fluxdetectionlimit import FluxDetectionLimit as fdl
 from diive.pkgs.echires.fluxdetectionlimit import FluxDetectionLimit as flux_detection_limit
@@ -111,6 +113,8 @@ __all__ = [
     'MeasurementOffsetFromReplicate',
     'remove_relativehumidity_offset',
     'remove_radiation_zero_offset',
+    'WindDirOffset',
+    'wind_dir_offset',
 
     # Packages: Variables
     'et_from_le',